made freelancer form reusable, added props, added freelancerhomepage
