+++ apps/freelance/src/translations/en/translation/description.ts
@@ -36,6 +36,7 @@ export default {
       mesTimeFrom: 'Please, input time from!',
       mesTimeTo: 'Please, input time to!',
       mesWork: 'Work history field is required!',
+      finish: 'Finish',
     },
     profileQp2: {
       pr_bar_completion_per: '50% completed',
@@ -55,6 +56,9 @@ export default {
     freelancerPageInfo: {
       sendOffer: 'Send offer',
       inviteInterview: 'Invite to interview',
+    },
+    freelancerEditProfile: {
+      save: 'Save',
     },
     router: {
       toPage2: 'Click here for page 2.',
