+++ libs/components/src/lib/constants/routes.ts
@@ -10,8 +10,8 @@ export const routes = {
   // Protected Routes
   role: '/role',
   //Freelancer Routes
-  profileQuestions1: '/profile-questions-1',
-  profileQuestions2: '/profile-questions-2',
+  profileQuestions: '/profile-questions',
+  freelancerProfile: '/:id/my-profile',
   findJobs: '/findjobs',
   jobDetails: '/jobdetails',
   proposalsList: '/job/:id/proposals',
@@ -20,7 +20,4 @@ export const routes = {
   jobPost: '/owner-profile/job-post',
   ownerProfile: '/owner-profile',
   talents: '/talents',
-};
-export const proFileQuestions = {
-  profileQuestions2: '/profile-questions-2',
 };
