+++ libs/components/tsconfig.json
@@ -13,7 +13,8 @@
     "noFallthroughCasesInSwitch": true,
     "baseUrl": ".",
     "paths": {
-      "src/*": ["../../apps/freelance/src/*"]
+      "src/*": ["../../apps/freelance/src/*"],
+      "@freelance/components*": ["../../libs/components/src/index.ts"]
     }
   },
   "files": [],
